# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Batra Butterfly Footwear Business OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock:
    get:
      operationId: listStock
      tags: [stock]
      summary: List stock items
      responses:
        "200":
          description: Stock list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockItem"
    post:
      operationId: createStockItem
      tags: [stock]
      summary: Create a stock product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockItemBody"
      responses:
        "201":
          description: Created stock item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockItem"
  /stock/analyze:
    post:
      operationId: analyzeStockImage
      tags: [stock]
      summary: Analyze uploaded image for article code suggestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeImageBody"
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageAnalysisResult"
  /stock/confirm:
    post:
      operationId: confirmStockMovement
      tags: [stock]
      summary: Confirm stock movement after image analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmStockBody"
      responses:
        "201":
          description: Stock movement confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
  /stock/analyze-sale:
    post:
      operationId: analyzeSaleImage
      tags: [stock]
      summary: Analyze a sale image to detect multiple sold items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeSaleImageBody"
      responses:
        "200":
          description: AI sale analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleAnalysisResult"
  /stock/confirm-sale:
    post:
      operationId: confirmSale
      tags: [stock]
      summary: Confirm a sale, deducting all detected items from stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmSaleBody"
      responses:
        "201":
          description: Sale confirmed with stock deductions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmSaleResult"
...
components:
  schemas:
...
    StockItem:
      type: object
      properties:
        id:
          type: integer
        articleCode:
          type: string
        name:
          type: string
        price:
          type: number
        purchasePrice:
          type: ["number", "null"]
        currentStock:
          type: integer
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - articleCode
        - name
        - price
        - currentStock
        - imageUrl
        - createdAt
    CreateStockItemBody:
      type: object
      properties:
        articleCode:
          type: string
        name:
          type: string
        price:
          type: number
        purchasePrice:
          type: number
        currentStock:
          type: integer
        imageUrl:
          type: string
      required:
        - articleCode
        - name
        - price
        - currentStock
    AnalyzeImageBody:
      type: object
      properties:
        imageBase64:
          type: string
        imageUrl:
          type: string
        imageBase64s:
          type: array
          items:
            type: string
          description: Multiple base64 images for better recognition
        imageUrls:
          type: array
          items:
            type: string
          description: Multiple image URLs for better recognition
      required: []
    ImageAnalysisResult:
      type: object
      properties:
        predictedArticleCode:
          type: string
        confidence:
          type: number
        suggestedName:
          type: string
        suggestedPrice:
          type: number
        matchedProduct:
          $ref: "#/components/schemas/StockItem"
        allProducts:
          type: array
          items:
            $ref: "#/components/schemas/StockItem"
        imageUrls:
          type: array
          items:
            type: string
          description: All uploaded image URLs/base64 strings used for analysis
      required:
        - predictedArticleCode
        - confidence
        - allProducts
    ConfirmStockBody:
      type: object
      properties:
        productId:
          type: integer
        articleCode:
          type: string
        name:
          type: string
        price:
          type: number
        purchasePrice:
          type: number
        type:
          type: string
          enum: [in, out]
        quantity:
          type: integer
        imageUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
          description: Multiple images to store for the product
        supplierId:
          type: integer
        supplierName:
          type: string
      required:
        - type
        - quantity
        - articleCode
    AnalyzeSaleImageBody:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image containing multiple sold items
      required:
        - imageBase64
    SaleDetectedItem:
      type: object
      properties:
        articleCode:
          type: string
        quantity:
          type: integer
        confidence:
          type: number
        matchedProduct:
          $ref: "#/components/schemas/StockItem"
        notFound:
          type: boolean
          description: True if the article code was not found in inventory
      required:
        - articleCode
        - quantity
        - confidence
    SaleAnalysisResult:
      type: object
      properties:
        detectedItems:
          type: array
          items:
            $ref: "#/components/schemas/SaleDetectedItem"
        allProducts:
          type: array
          items:
            $ref: "#/components/schemas/StockItem"
        imageUrl:
          type: string
      required:
        - detectedItems
        - allProducts
    ConfirmSaleBody:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              articleCode:
                type: string
              productId:
                type: integer
              quantity:
                type: integer
            required:
              - articleCode
              - quantity
        imageUrl:
          type: string
      required:
        - items
    ConfirmSaleResult:
      type: object
      properties:
        movements:
          type: array
          items:
            $ref: "#/components/schemas/StockMovement"
        totalItems:
          type: integer
        message:
          type: string
      required:
        - movements
        - totalItems
        - message
...
    StockMovement:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        articleCode:
          type: string
        type:
          type: string
          enum: [in, out]
        quantity:
          type: integer
        imageUrl:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - articleCode
        - type
        - quantity
        - imageUrl
        - date
        - createdAt
...